% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxcox.R
\name{bc}
\alias{bc}
\title{Apply the boxcox transformation to an univariate time-serie.}
\usage{
bc(x, lambda)
}
\arguments{
\item{x}{a vector of values.}

\item{lambda}{the scalar parameter of the boxcox transformation.}
}
\value{
returns a vector of transformed values.
}
\description{
Apply the boxcox transformation to an univariate time-serie.
}
\details{
The univarite time-serie must be positive
}

