% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_fit.R
\name{compute_par}
\alias{compute_par}
\alias{compute_par.gauss_fit}
\alias{compute_par.gev_fit}
\alias{compute_par.gpd_fit}
\title{compute_par generic}
\usage{
compute_par(object, newdata)

\method{compute_par}{gpd_fit}(object, newdata)

\method{compute_par}{gev_fit}(object, newdata)

\method{compute_par}{gauss_fit}(object, newdata)
}
\arguments{
\item{object}{object of class gauss_fit, gev_fit or gpd_fit.}

\item{newdata}{data.frame giving the necessary information to compute the parameters of the fitted models at different times.}
}
\description{
returns the parameters of the fitted model given the information provided in newdata
}
\section{Methods (by class)}{
\itemize{
\item \code{gpd_fit}: returns the parameters of the fitted GPD at different times.

\item \code{gev_fit}: returns the parameters of the fitted GEV at different times.

\item \code{gauss_fit}: returns the parameters of the fitted Gaussian at different times.
}}

