% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARg-package.r
\docType{package}
\name{FARg}
\alias{FARg}
\alias{FARg-package}
\title{FARg}
\description{
The package provides tools to compute the Fraction of Attributable Risk from  a gaussian, a GEV or a GPD time-serie.
}
\details{
To compute the FAR, one first needs to fit the time serie using either gauss_fit, gev_fit or gpd_fit. Then the function get_far can be used to estimate the FAR and the function boot_far or prof_far can be used to estimate confidences intervals.
}
