% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FARg.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{FARg : the package provides tools to compute the Fraction of Attributable Risk from  a gaussian, a GEV or a GPD time serie.}
\description{
To compute the FAR, one first needs to fit the time serie using either gauss_fit, gev_fit or gpd_fit. Then the function get_far can be used to estimate the FAR and the function boot_ic or prof_ic can be used to estimate confidences intervals.
}

