% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_far.R
\name{get_p}
\alias{get_p}
\alias{get_p.gauss_fit}
\alias{get_p.gpd_fit}
\alias{get_p.gev_fit}
\title{get_p generic.}
\usage{
get_p(object, ...)

\method{get_p}{gauss_fit}(object, pnt, ...)

\method{get_p}{gpd_fit}(object, pnt, under_threshold = FALSE, ...)

\method{get_p}{gev_fit}(object, pnt, ...)
}
\arguments{
\item{object}{an object of class gauss_fit, gev_fit, gpd_fit.}

\item{...}{Arguments to be passed to methods,}

\item{pnt}{a point which consists of a line of data.frame containing the same variables used for the fit. It can bet set using the function \code{set_pnt}}

\item{under_threshold}{used for gpd_fit. It tells whether it provides an estimated when the point is under the GPD threshold. In which case, gives the empirical probability obtained by quantile regression.}
}
\description{
Compute the probability of exeeding a given point.
}
\section{Methods (by class)}{
\itemize{
\item \code{gauss_fit}: for a gaussian fit.

\item \code{gpd_fit}: for a GPD fit.

\item \code{gev_fit}: for a GEV fit.
}}

\examples{
data(tas)

ge_fit <- gev_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")
gp_fit <- gpd_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year", qthreshold=0.9)
ga_fit <- gauss_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")

t1 <- 2003
t0 <- 1990
xp <- 1.6
pnt1 <- set_pnt(t1, xp, time_var="year", tas)
pnt0 <- set_pnt(t0, xp, time_var="year", tas)

get_p(ga_fit ,pnt1)
get_far(ga_fit, pnt0, pnt1)

get_p(gp_fit, pnt1)
get_far(gp_fit, pnt0, pnt1)

get_p(ge_fit, pnt1)
get_far(ge_fit, pnt0, pnt1)
}
