% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_q.R
\name{get_q}
\alias{get_q}
\alias{get_q.gauss_fit}
\alias{get_q.gev_fit}
\alias{get_q.gpd_fit}
\title{get_q generic.}
\usage{
get_q(object, ...)

\method{get_q}{gauss_fit}(object, pnt, ...)

\method{get_q}{gpd_fit}(object, pnt, ...)

\method{get_q}{gev_fit}(object, pnt, ...)
}
\arguments{
\item{object}{an object of class gauss_fit, gev_fit, gpd_fit.}

\item{...}{Arguments to be passed to methods,}

\item{pnt}{a point which consists of a line of data.frame containing the same variables used for the fit. It can bet set using the function \code{set_pnt}}
}
\description{
Compute the quantile corresponding to the probability of exceedance p.
}
\section{Methods (by class)}{
\itemize{
\item \code{gauss_fit}: for a Gaussian fit.

\item \code{gpd_fit}: for a GPD fit.

\item \code{gev_fit}: for a GEV fit.
}}
\examples{
data(tas)

ge_fit <- gev_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")
gp_fit <- gpd_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year", qthreshold=0.9)
ga_fit <- gauss_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")

t1 <- 2003
t0 <- 1990
p <- 0.01
pnt1 <- set_pnt(t1, p, time_var="year", tas)
pnt0 <- set_pnt(t0, p, time_var="year", tas)

get_q(ga_fit ,pnt1)

get_q(gp_fit, pnt1)

get_q(ge_fit, pnt1)
}

