% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_fit.R
\name{gev_fit}
\alias{gev_fit}
\title{Fit a time serie with a GEV distribution  .}
\usage{
gev_fit(y, data, mu_mod = ~1, sig_mod = ~1, time_var, init = NULL)
}
\arguments{
\item{y}{the time serie to be fitted.}

\item{data}{a data.frame object with  where the function looks first for the variables y, time_var and the covariates specified in the mu_mod and sig_mod arguments.}

\item{mu_mod}{a formula defining the covariates the location parameter of the GEV depends linearly on.}

\item{sig_mod}{a formula defining the covariates the scale parameter of the GEV depends linearly on.}

\item{time_var}{a variable used to define the time in the time serie. It can also be a string giving the variable name.}

\item{init}{vector of initialization parameter for the minimization of the negative log-likelihood. if NULL, the initialisation is done using the function fevd from the extRemes packages.}
}
\value{
returns an object of class gev_fit. It contains the nlminb output which provides the estimated parameters as well the minimum of the negative log-likelihood. The arguments use to call gev_fit.
}
\description{
\code{gev_fit} fits a time serie with a GEV distribution a where the location and scale parameters depend linearly on a set of covariates.
}
\details{
MLE fit of a time serie y using a GEV distribution  where the location and the scale parameters depend linearly on a set of covariates. The optimization of the negative log-likelihood is done with the nlminb function in R.
}
\examples{
data(tas)
#Example with the same covariate for the mean and variance parameter
ge_fit <- gev_fit(eur_tas, data=tas, mu_mod=~avg_gbl_tas, sig_mod=~avg_gbl_tas, time_var="year")
# get the values of the mean and variance parameters of the GEV at each time
compute_par(ge_fit, tas)
# plot diagnostic plot of the fit : standardized residuals plots, qqplot, density of fitted vs theorical density, times series ans return levels
plot(ge_fit)
}

