% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_fit.R
\name{plot.gauss_fit}
\alias{plot.gauss_fit}
\title{Plot diagnostics for an gauss_fit object}
\usage{
\method{plot}{gauss_fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{gauss_fit}.}

\item{...}{Arguments to be passed to methods,}
}
\description{
Diagnostic plots of a gaussian fit : standardized residuals plots, qqplot, density of fitted vs theorical density, times series ans return levels
}
\details{
Diagnostic plots of the fit : standardized residuals plots, qqplot, density of fitted vs theorical density, times series ans return levels. This function is just a wrapper of \code{plot.fevd} from the extRemes packages : the gaussian is fitted again with the fevd function with initial parameters taken from the \code{gauss_fit} object and then plotted using \code{plot.fevd}.
}
\examples{
data(tas)
#Example with the same covariate for the mean and variance parameter
ga_fit <- gauss_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")
# get the values of the mean and variance parameters of the gaussian at each time
compute_par(ga_fit, tas)
# plot diagnostic plot of the fit : standardized residuals plots, qqplot, density of fitted vs theorical density, times series ans return levels
plot(ga_fit)
}

