% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{plot.gev_fit}
\alias{plot.gev_fit}
\title{Plot diagnostics for an gev_fit object}
\usage{
\method{plot}{gev_fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{gev_fit}.}

\item{...}{Arguments to be passed to methods,}
}
\description{
Diagnostic plots of a GEV fit : standardized residuals plots, qqplot, density of fitted vs theorical density, times series ans return levels
}
\details{
Diagnostic plots of the fit : standardized residuals plots, qqplot, density of fitted vs theorical density, times series ans return levels. This function is just a wrapper of \code{plot.fevd} from the extRemes packages : the GEV is fitted again with the fevd function with initial parameters taken from the \code{gev_fit} object and then plotted using \code{plot.fevd}.
}
\examples{
data(tas)
#Example with the same covariate for the mean and variance parameter
ge_fit <- gev_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")
# get the values of the mean and variance parameters of the GEV at each time
compute_par(ge_fit, tas)
# plot diagnostic plot of the fit : standardized residuals plots, qqplot, density of fitted vs theorical density, times series ans return levels
plot(ge_fit)
}
