% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{plot.gpd_fit}
\alias{plot.gpd_fit}
\title{Plot diagnostics for an gpd_fit object}
\usage{
\method{plot}{gpd_fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{gpd_fit}.}

\item{...}{Arguments to be passed to methods,}
}
\description{
Diagnostic plots of a GPD fit : standardized residuals plots, qqplot, density of fitted vs theorical density, times series ans return levels
}
\details{
Diagnostic plots of the fit : qqplot, density of fitted vs theorical density, times series ans return levels. This function is just a wrapper of \code{plot.fevd} from the extRemes packages : the GPD is fitted again with the fevd function with initial parameters taken from the \code{gpd_fit} object and then plotted using \code{plot.fevd}.
}
\examples{
data(tas)
#Example with the same covariate for the mean and variance parameter
gp_fit <- gpd_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year", qthreshold=0.9)
# get the values of the mean and variance parameters of the GEV at each time
compute_par(gp_fit, tas)
# plot diagnostic plot of the fit : qqplot, density of fitted vs theorical density, times series ans return levels
plot(gp_fit)
}
