% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{print.gpd_fit}
\alias{print.gpd_fit}
\alias{print.gauss_fit}
\alias{print.gev_fit}
\title{Print the optimisation results of the fit.}
\usage{
\method{print}{gpd_fit}(x, ...)

\method{print}{gauss_fit}(x, ...)

\method{print}{gev_fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{gauss_fit}, \code{gev_fit} or \code{gev_fit}.}

\item{...}{Arguments to be passed to methods,}
}
\description{
Print the results of the optimization function \code{nlminb} used to fit the model.
}
\examples{
ga_fit <- gauss_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")
print(ga_fit)
ge_fit <- gev_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")
print(ge_fit)
gp_fit <- gpd_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year", qthreshold=0.9)
print(gp_fit)
}
