% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tcplot.R
\name{select_thresh}
\alias{select_thresh}
\title{Select threshold from tcplot output.}
\usage{
select_thresh(tc)
}
\arguments{
\item{tc}{result from the tcplot function of the evmix packages}
}
\value{
returns the value of the threshold selected
}
\description{
Select trheshold from threshold stability plot from the tcplot function of the evmix package.
}
\details{
The threhshold is selected such as as the minimum threshold for which the transformed parameters of the GPD model fitted are within the confidence intervals for every higher threshold.
}

