% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_far.R
\name{set_pnt}
\alias{set_pnt}
\title{Create a point.}
\usage{
set_pnt(time, y, time_var = "", data = NULL)
}
\arguments{
\item{time}{the time we want the point to be closest to.}

\item{y}{the value we want to set the point at the given time.}

\item{time_var}{the variable used as time in data}

\item{data}{the data.frame containing the relevent information to create the point}
}
\description{
Create a point at a given time from a given data.
}
\details{
The function extract the point of the data set, time of which is the closest to the one specified.
}
\examples{
data(tas)

ge_fit <- gev_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")
gp_fit <- gpd_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year", qthreshold=0.9)
ga_fit <- gauss_fit(eur_tas, data=tas, mu_mod=~gbl_tas, sig_mod=~gbl_tas, time_var="year")

t1 <- 2003
t0 <- 1990
xp <- 1.6
pnt1 <- set_pnt(t1, xp, time_var="year", tas)
pnt0 <- set_pnt(t0, xp, time_var="year", tas)

get_p(ga_fit ,pnt1)
get_far(ga_fit, pnt0, pnt1)

get_p(gp_fit, pnt1)
get_far(gp_fit, pnt0, pnt1)

get_p(ge_fit, pnt1)
get_far(ge_fit, pnt0, pnt1)
}
