% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize the times series.}
\usage{
standardize(y, data, mu_mod = ~1, sig2_mod = ~1)
}
\arguments{
\item{y}{the time series values we want to remove the average trend and the heteroscedasticity from.}

\item{data}{a data.frame with the covariates needed to estimates the time series trend and variance.}

\item{mu_mod}{a formula giving the covariates the time-serie trend depends linearly on.}

\item{sig2_mod}{a formula giving the covariates the time-serie variance depends linearly on.}
}
\value{
returns a list with the arguments of the function as well as :
\itemize{
\item y_std, the standardized time serie.
\item mu_fit, the linear model fitted for the trend.
\item sig2_fit, the linear model fitted for the variance.
}
}
\description{
Trys to remove the trend and the heteroscedaticity of the time series estimated by linears models.
}
\details{
Compute the linear trend of the time series as well as  the evolution of the variance residuals with respect to the specified covariates and then standardize the time series accordingly.
}
\examples{
data(tas)
y_std <- standardize(eur_tas, data=tas, mu_mod=~avg_gbl_tas, sig2_mod=~avg_gbl_tas)
y_std_fit <- gpd_fit(y_std$y_std ,  data=tas, time_var="year", qthreshold=0.8)
t1 <- 2003
t0 <- 1990
xp <- 1.6
pnt1 <- set_pnt(t1, xp, time_var="year", tas)
pnt0 <- set_pnt(t0, xp, time_var="year", tas)
get_far(y_std, y_std_fit, pnt0, pnt1)
boot_ic(y_std,  y_std_fit, xp, t0, t1)
}

