% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tas.R
\name{tas}
\alias{tas}
\title{European and global summer mean temperature anomalies from 1850 to 2100"}
\format{A data frame with 1255 rows and 333 variables}
\description{
European and global summer temperature anomalies from 5 runs of the CNRM-CM5 climate models. Summer corresponds to the months JJA. Anomalies are computed from a climatology from 1961 to 1990. First part of the simulations are historical runs and afterwards the scenario RCP8.5 is used.
}
\details{
\itemize{
  \item year. year of the anomalies
  \item eur_tas. European summer temperature anomalies (K)
  \item avg_gbl_tas. Global summer temperature anomalies (K)
}
}

